% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleAnswer.R
\name{getPermutations}
\alias{getPermutations}
\title{Title}
\usage{
getPermutations(num_choice, answer_hash)
}
\arguments{
\item{num_choice}{number of choices you have in the question}

\item{answer_hash}{the hash number of your right answer, it should be shown to you on the screen if you get it wrong at the first time.}
}
\value{
a string indicating that function is complete.
}
\description{
Title
}
\examples{
#example code
getPermutations(6, "<test hash>")

\dontrun{
getPermutations(2048)
}
}
